requirements.txt
Flask==3.1.0
flask-cors==5.0.1
google-analytics-data==0.18.18
google-auth==2.39.0
langgraph==0.3.31
pandas==2.2.3
python-dotenv==1.0.0
requests==2.32.3
gunicorn==21.2.0